First draft of overview
This first draft of the case overview is just a plaintext, un-cited version which I typed out this morning. The sources for everything mentioned here can be found somewhere on my Twitter stream from the last few days. I will be formatting it to look nice & adding citations to these claims whenever I have the time.