Added new section "Post-Case Investigation"
Since the case has been closed, I thought it was better to have all post-case information in a new section. I may move this to the bottom of the overview (after the legal implications section) since it would flow better chronologically. I would first want to go through the rest of the overview and transition it to past-tense.